Add files via upload
This reads in URLs of the congressional appropriation's subcommittee of agriculture's annual reports from 2000-2022. It removes all tables and numbers from the text and categorizes the text by agency provision creating a CSV with 2 columns. The first column is the agency name, and the second is the text describing their annual provision.